# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/overview": {
      "get": {
        "operationId": "AdminController_overview_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStatsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Platform-wide counts, uptime and DB latency",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/v1/admin/events": {
      "get": {
        "operationId": "AdminController_listEvents_v1",
        "parameters": [
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DEBUG",
                "INFO",
                "WARN",
                "ERROR",
                "FATAL"
              ]
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "schema": {
              "example": "api",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text match on message or event code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "required": false,
            "in": "query",
            "description": "ISO timestamp lower bound",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "afterId",
            "required": false,
            "in": "query",
            "description": "Return only events with id greater than this — the live-tail cursor",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemEventPageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Paginated operational event log",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/events/stream": {
      "get": {
        "operationId": "AdminController_streamEvents_v1",
        "parameters": [
          {
            "name": "afterId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Live event feed (SSE)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/jobs": {
      "get": {
        "description": "A scheduler nobody can inspect is a scheduler nobody trusts. The case that matters is FAILURE: a purge that has been erroring for a week looks exactly like a purge with nothing to do unless the run log is visible.",
        "operationId": "AdminController_scheduledJobs_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledJobsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Scheduled jobs and their last run outcome",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/tracing": {
      "get": {
        "operationId": "AdminController_tracingStatus_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Whether the request-trace window is open, and until when",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "description": "Records EVERY request to system_events so one user's request can be followed end to end. Expires on its own — a trace that must be switched off by hand becomes a permanent activity log of every user, which DEC-016 does not permit an operator to keep.",
        "operationId": "AdminController_enableTracing_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableTracingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Open a time-boxed request-trace window",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_disableTracing_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Close the trace window early",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/events/prune": {
      "post": {
        "operationId": "AdminController_pruneEvents_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Drop events past the retention window",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/audit": {
      "get": {
        "operationId": "AdminController_listAudit_v1",
        "parameters": [
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventPageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Global audit trail",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "description": "Platform metadata and activity COUNTS only. This endpoint cannot return a balance, amount, payee or note — operators have no access to any user's financial rows (DEC-002 #12, DEC-007).",
        "operationId": "AdminController_listUsers_v1",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches email or display name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "SUSPENDED",
                "PENDING_DELETION"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata and counts; never financial data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserPageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "User metadata list",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{userId}": {
      "get": {
        "operationId": "AdminController_getUser_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Single user metadata",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "description": "Allowlist: display_name, locale, timezone. is_super_admin is deliberately not editable here — promotion is a service-role SQL operation (DEC-013).",
        "operationId": "AdminController_updateUser_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update the operator-editable subset of a profile",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{userId}/suspend": {
      "post": {
        "operationId": "AdminController_suspendUser_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Suspend an account (Auth ban + status)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/users/{userId}/reinstate": {
      "post": {
        "operationId": "AdminController_reinstateUser_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Lift a suspension, and cancel any pending deletion",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/purge": {
      "post": {
        "description": "Irreversible. Removes application data in FK dependency order, then the auth user. Only touches accounts whose 30-day grace has already expired.",
        "operationId": "AdminController_runPurge_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Run the deletion purge for expired grace periods",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/settings": {
      "get": {
        "operationId": "AdminController_listSettings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppSettingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "All global settings, public and private",
        "tags": [
          "Admin"
        ]
      },
      "put": {
        "operationId": "AdminController_updateSettings_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Update one or more known settings",
        "tags": [
          "Admin"
        ]
      }
    },
    "/v1/admin/releases": {
      "get": {
        "operationId": "AdminController_getReleases_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileReleaseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "supabase-auth": []
          }
        ],
        "summary": "Current operator-managed mobile release",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "NoorixFin API",
    "description": "Personal and household finance management API. All financial amounts are minor-unit decimal strings. All timestamps are ISO 8601 UTC. Every failure body matches the `ApiErrorBody` schema; branch on its `code`, never on `message`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "supabase-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase access token"
      }
    },
    "schemas": {
...
      "SystemEventPageResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemEventResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "ScheduledJobsResponseDto": {
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledJobResponseDto"
            }
          },
          "run_at": {
            "type": "string"
          }
        },
        "required": [
          "jobs",
          "run_at"
        ]
      },
...
    }
  }
}